Adding secureboot option to build optionrom (#24050)
Fix up ionic_unmap_bars()

Checking ionic_close() and ionic_remove()

Size_tversus__le16typemismatchinlengthvaluein ionic_transmit

+++ platform/drivers/pxe/ionic.h
@@ -276,7 +276,7 @@ int ionic_lif_rx_mode(struct lif *lif, unsigned int rx_mode);
 int ionic_lif_quiesce(struct lif *lif);
 void ionic_rx_flush(struct lif *lif);
 void ionic_tx_flush(struct net_device *netdev, struct lif *lif);
-void ionic_rx_fill(struct net_device *netdev, int length);
+void ionic_rx_fill(struct net_device *netdev, u16 length);
 void ionic_poll_rx(struct net_device *netdev);
 void ionic_poll_tx(struct net_device *netdev);
 bool ionic_q_has_space(struct queue *q, unsigned int want);
